Contraction

Contraction path optimization and execution is delegated to the EinExprs library. A EinExpr is a lower-level form of a Tensor Network, in which the contraction path has been laid out as a tree. It is similar to a symbolic expression (i.e. Expr) but in which every node represents an Einstein summation expression (aka einsum).

EinExprs.einexpr — Method

einexpr(tn::TensorNetwork; optimizer = EinExprs.Greedy, output = inds(tn, :open), kwargs...)

Search a contraction path for the given TensorNetwork and return it as a EinExpr.

Keyword Arguments

optimizer Contraction path optimizer. Check EinExprs documentation for more info.

outputs Indices that won't be contracted. Defaults to open indices.

kwargs Options to be passed to the optimizer.

See also: contract.

Tenet.contract — Function

contract(a::Tensor[, b::Tensor, dims=nonunique([inds(a)..., inds(b)...])])

Perform tensor contraction operation.

contract(tn::TensorNetwork; kwargs...)

Contract a TensorNetwork. The contraction order will be first computed by einexpr.

The kwargs will be passed down to the einexpr function.

See also: einexpr, contract!.

Tenet.contract! — Function

contract!(tn::TensorNetwork, index::Symbol)

In-place contraction of tensors connected to index.

See also: contract.