Quantum Tensor Networks

Tenet.Quantum — Type

Quantum

Tensor Network with a notion of "causality". This leads to the notion of sites and directionality (input/output).

Notes

Indices are referenced by Sites.

Tenet.TensorNetwork — Method

TensorNetwork(q::Quantum)

Returns the underlying TensorNetwork of a Quantum Tensor Network.

Base.adjoint — Method

adjoint(q::Quantum)

Returns the adjoint of a Quantum Tensor Network; i.e. the conjugate Tensor Network with the inputs and outputs swapped.

Tenet.sites — Function

sites(q::Quantum)

Returns the sites of a Quantum Tensor Network.

Tenet.nsites — Function

nsites(q::Quantum)

Returns the number of sites of a Quantum Tensor Network.

Queries

EinExprs.inds — Method

inds(tn::Quantum, set::Symbol = :all, args...; kwargs...)

Options:

:at: index at a site

Tenet.tensors — Method

tensors(tn::Quantum, query::Symbol, args...; kwargs...)

Options:

:at: tensor at a site

Connecting Quantum Tensor Networks

Tenet.inputs — Function

inputs(q::Quantum)

Returns the input sites of a Quantum Tensor Network.

Tenet.outputs — Function

outputs(q::Quantum)

Returns the output sites of a Quantum Tensor Network.

Missing docstring.

Missing docstring for lanes. Check Documenter's build log for details.

Tenet.ninputs — Function

ninputs(q::Quantum)

Returns the number of input sites of a Quantum Tensor Network.

Tenet.noutputs — Function

noutputs(q::Quantum)

Returns the number of output sites of a Quantum Tensor Network.

Missing docstring.

Missing docstring for nlanes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Socket. Check Documenter's build log for details.

Tenet.socket — Method

socket(q::Quantum)

Returns the socket of a Quantum Tensor Network; i.e. whether it is a Scalar, State or Operator.

Tenet.Scalar — Type

Scalar <: Socket

Socket representing a scalar; i.e. a Tensor Network with no open sites.

Tenet.State — Type

State <: Socket

Socket representing a state; i.e. a Tensor Network with only input sites (or only output sites if dual = true).

Tenet.Operator — Type

Operator <: Socket

Socket representing an operator; i.e. a Tensor Network with both input and output sites.

Base.merge — Method

merge(a::Quantum, b::Quantum...)

Merges multiple Quantum Tensor Networks into a single one by connecting input/output sites.