Projected Entangled Pair States (PEPS)

Projected Entangled Pair States (PEPS) are a Quantum Tensor Network ansatz whose tensors are laid out in a 2D lattice. Depending on the boundary conditions, the chains can be open or closed (i.e. periodic boundary conditions).

fig = Figure() # hide

tn_open = rand(PEPS{Open}, rows=10, cols=10, χ=4) # hide
tn_periodic = rand(PEPS{Periodic}, rows=10, cols=10, χ=4) # hide

plot!(fig[1,1], tn_open, layout=Stress(seed=1)) # hide
plot!(fig[1,2], tn_periodic, layout=Stress(seed=10,dim=2,iterations=100000)) # hide

Label(fig[1,1, Bottom()], "Open") # hide
Label(fig[1,2, Bottom()], "Periodic") # hide

fig # hide

Projected Entangled Pair Operators (PEPO)

fig = Figure() # hide

tn_open = rand(PEPO{Open}, rows=10, cols=10, χ=4) # hide
tn_periodic = rand(PEPO{Periodic}, rows=10, cols=10, χ=4) # hide

plot!(fig[1,1], tn_open, layout=Stress(seed=1)) # hide
plot!(fig[1,2], tn_periodic, layout=Stress(seed=10,dim=2,iterations=100000)) # hide

Label(fig[1,1, Bottom()], "Open") # hide
Label(fig[1,2, Bottom()], "Periodic") # hide

fig # hide

Tenet.ProjectedEntangledPair — Type

ProjectedEntangledPair{P<:Plug,B<:Boundary} <: Quantum

A generic ansatz representing Projected Entangled Pair States (PEPS) and Projected Entangled Pair Operators (PEPO). Type variable P represents the Plug type (State or Operator) and B represents the Boundary type (Open or Periodic).

Ansatz Fields

χ::Union{Nothing,Int} Maximum virtual bond dimension.

Tenet.ProjectedEntangledPair — Method

ProjectedEntangledPair{P,B}(arrays::Matrix{AbstractArray}; χ::Union{Nothing,Int} = nothing, order = defaultorder(ProjectedEntangledPair{P}))

Construct a TensorNetwork with ProjectedEntangledPair ansatz, from the arrays of the tensors.

Keyword Arguments

χ Maximum virtual bond dimension. Defaults to nothing.

order Order of the tensor indices on arrays. Defaults to (:l, :r, :u, :d, :o) if P is a State, (:l, :r, :u, :d, :i, :o) if Operator.