Transformations

In tensor network computations, it is good practice to apply various transformations to simplify the network structure, reduce computational cost, or prepare the network for further operations. These transformations modify the network's structure locally by permuting, contracting, factoring or truncating tensors.

A crucial reason why these methods are indispensable lies in their ability to drastically reduce the problem size of the contraction path search and also the contraction. This doesn't necessarily involve reducing the maximum rank of the Tensor Network itself, but more importantly, it reduces the size (or rank) of the involved tensors.

Our approach is based in (Gray and Kourtis, 2021), which can also be found in quimb.

In Tenet, we provide a set of predefined transformations which you can apply to your TensorNetwork using both the transform/transform! functions.

Tenet.transform — Function

transform(tn::TensorNetwork, config::Transformation)
transform(tn::TensorNetwork, configs)

Return a new TensorNetwork where some Transformation has been performed into it.

See also: transform!.

Tenet.transform! — Function

transform!(tn::TensorNetwork, config::Transformation)
transform!(tn::TensorNetwork, configs)

In-place version of transform.

Available transformations

Hyperindex converter

Tenet.HyperindConverter — Type

HyperindConverter <: Transformation

Convert hyperindices to COPY-tensors, represented by DeltaArrays. This transformation is always used by default when visualizing a TensorNetwork with plot.

Diagonal reduction

Tenet.DiagonalReduction — Type

DiagonalReduction <: Transformation

Reduce the dimension of a Tensor in a TensorNetwork when it has a pair of indices that fulfil a diagonal structure.

Keyword Arguments

atol Absolute tolerance. Defaults to 1e-12.

set_theme!(resolution=(800,200)) # hide
fig = Figure() #hide

data = zeros(Float64, 2, 2, 2, 2) #hide
for i in 1:2 #hide
    for j in 1:2 #hide
        for k in 1:2 #hide
            data[i, i, j, k] = k #hide
        end #hide
    end #hide
end #hide

A = Tensor(data, (:i, :j, :k, :l)) #hide
B = Tensor(rand(2, 2), (:i, :m)) #hide
C = Tensor(rand(2, 2), (:j, :n)) #hide

tn = TensorNetwork([A, B, C]) #hide
reduced = transform(tn, Tenet.DiagonalReduction) #hide

smooth_annotation!( #hide
    fig[1, 1]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = -0.21, #hide
    offset_y = -0.42, #hide
    radius_x = 0.38, #hide
    radius_y = 0.8, #hide
    num_waves = 6, #hide
    fluctuation_amplitude = 0.02, #hide
    phase_shift = 0.0) #hide
plot!(fig[1, 1], tn, layout=Spring(iterations=1000, C=0.5, seed=100); node_color=[red, orange, orange]) #hide

smooth_annotation!( #hide
    fig[1, 2]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = 0.1, #hide
    offset_y = -0.35, #hide
    radius_x = 0.38, #hide
    radius_y = 1.1, #hide
    num_waves = 5, #hide
    fluctuation_amplitude = 0.02, #hide
    phase_shift = 1.9) #hide
plot!(fig[1, 2], reduced, layout=Spring(iterations=1000, C=0.5, seed=100),  node_color=[orange, orange, red, :black]) #hide

Label(fig[1, 1, Bottom()], "Original") #hide
Label(fig[1, 2, Bottom()], "Transformed") #hide

fig #hide

Anti-diagonal reduction

Tenet.AntiDiagonalGauging — Type

AntiDiagonalGauging <: Transformation

Reverse the order of tensor indices that fulfill the anti-diagonal condition. While this transformation doesn't directly enhance computational efficiency, it sets up the TensorNetwork for other operations that do.

Keyword Arguments

atol Absolute tolerance. Defaults to 1e-12.

skip List of indices to skip. Defaults to [].

Rank simplification

Tenet.RankSimplification — Type

RankSimplification <: Transformation

Preemptively contract tensors whose result doesn't increase in size.

set_theme!(resolution=(800,200)) # hide
fig = Figure() #hide

A = Tensor(rand(2, 2, 2, 2), (:i, :j, :k, :l)) #hide
B = Tensor(rand(2, 2), (:i, :m)) #hide
C = Tensor(rand(2, 2, 2), (:m, :n, :o)) #hide
E = Tensor(rand(2, 2, 2, 2), (:o, :p, :q, :j)) #hide

tn = TensorNetwork([A, B, C, E]) #hide
reduced = transform(tn, Tenet.RankSimplification) #hide

smooth_annotation!( #hide
    fig[1, 1]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = -0.32, #hide
    offset_y = -0.5, #hide
    radius_x = 0.25, #hide
    radius_y = 0.94, #hide
    num_waves = 6, #hide
    fluctuation_amplitude = 0.01, #hide
    phase_shift = 0.0) #hide
plot!(fig[1, 1], tn, layout=Spring(iterations=1000, C=0.5, seed=20); node_color=[orange, red, orange, orange]) #hide

smooth_annotation!( #hide
    fig[1, 2]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = 0.12, #hide
    offset_y = -0.62, #hide
    radius_x = 0.18, #hide
    radius_y = 0.46, #hide
    num_waves = 5, #hide
    fluctuation_amplitude = 0.01, #hide
    phase_shift = 0) #hide
plot!(fig[1, 2], reduced, layout=Spring(iterations=1000, C=0.5, seed=1); node_color=[red, orange, orange]) #hide

Label(fig[1, 1, Bottom()], "Original") #hide
Label(fig[1, 2, Bottom()], "Transformed") #hide

fig #hide

Column reduction

Tenet.ColumnReduction — Type

ColumnReduction <: Transformation

Truncate the dimension of a Tensor in a TensorNetwork when it contains columns with all elements smaller than atol.

Keyword Arguments

atol Absolute tolerance. Defaults to 1e-12.

skip List of indices to skip. Defaults to [].

set_theme!(resolution=(800,200)) # hide
fig = Figure() #hide

data = rand(3, 3, 3) #hide
data[:, 1:2, :] .= 0 #hide

A = Tensor(data, (:i, :j, :k)) #hide
B = Tensor(rand(3, 3), (:j, :l)) #hide
C = Tensor(rand(3, 3), (:l, :m)) #hide

tn = TensorNetwork([A, B, C]) #hide
reduced = transform(tn, Tenet.ColumnReduction) #hide

smooth_annotation!( #hide
    fig[1, 1]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = -1.12, #hide
    offset_y = -0.22, #hide
    radius_x = 0.35, #hide
    radius_y = 0.84, #hide
    num_waves = 4, #hide
    fluctuation_amplitude = 0.02, #hide
    phase_shift = 0.0) #hide
plot!(fig[1, 1], tn, layout=Spring(iterations=1000, C=0.5, seed=6); node_color=[red, orange, orange]) #hide

smooth_annotation!( #hide
    fig[1, 2]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = -0.64, #hide
    offset_y = 1.2, #hide
    radius_x = 0.32, #hide
    radius_y = 0.78, #hide
    num_waves = 5, #hide
    fluctuation_amplitude = 0.02, #hide
    phase_shift = 0) #hide

Label(fig[1, 1, Bottom()], "Original") #hide
Label(fig[1, 2, Bottom()], "Transformed") #hide
plot!(fig[1, 2], reduced, layout=Spring(iterations=2000, C=40, seed=8); node_color=[red, orange, orange]) #hide

fig #hide

Split simplification

Tenet.SplitSimplification — Type

SplitSimplification <: Transformation

Reduce the rank of tensors in the TensorNetwork by decomposing them using the Singular Value Decomposition (SVD). Tensors whose factorization do not increase the maximum rank of the network are left decomposed.

Keyword Arguments

atol Absolute tolerance. Defaults to 1e-10.

set_theme!(resolution=(800,200)) # hide
fig = Figure() #hide

v1 = Tensor([1, 2, 3], (:i,)) #hide
v2 = Tensor([4, 5, 6], (:j,)) #hide
m1 = Tensor(rand(3, 3), (:k, :l)) #hide

t1 = contract(v1, v2) #hide
tensor = contract(t1, m1)  #hide

tn = TensorNetwork([tensor, Tensor(rand(3, 3, 3), (:k, :m, :n)), Tensor(rand(3, 3, 3), (:l, :n, :o))]) #hide
reduced = transform(tn, Tenet.SplitSimplification) #hide

smooth_annotation!( #hide
    fig[1, 1]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = 0.24, #hide
    offset_y = 0.6, #hide
    radius_x = 0.32, #hide
    radius_y = 0.78, #hide
    num_waves = 5, #hide
    fluctuation_amplitude = 0.015, #hide
    phase_shift = 0.0) #hide
plot!(fig[1, 1], tn, layout=Spring(iterations=10000, C=0.5, seed=12); node_color=[red, orange, orange]) #hide

smooth_annotation!( #hide
    fig[1, 2]; #hide
    color = bg_blue, #hide
    xlims = [-2, 2], #hide
    ylims = [-2, 2], #hide
    offset_x = -0.2, #hide
    offset_y = -0.4, #hide
    radius_x = 1.1, #hide
    radius_y = 0.75, #hide
    num_waves = 3, #hide
    fluctuation_amplitude = 0.18, #hide
    phase_shift = 0.8) #hide

Label(fig[1, 1, Bottom()], "Original") #hide
Label(fig[1, 2, Bottom()], "Transformed") #hide
plot!(fig[1, 2], reduced, layout=Spring(iterations=10000, C=13, seed=151); node_color=[orange, orange, red, red, red]) #hide

fig #hide