Contraction

Contraction path optimization and execution is delegated to the EinExprs library. A EinExpr is a lower-level form of a Tensor Network, in which the contraction path has been laid out as a tree. It is similar to a symbolic expression (i.e. Expr) but in which every node represents an Einstein summation expression (aka einsum).

EinExprs.einexpr — Method

einexpr(tn::AbstractTensorNetwork; optimizer = EinExprs.Greedy, output = inds(tn, :open), kwargs...)

Search a contraction path for the given AbstractTensorNetwork and return it as a EinExpr.

Keyword Arguments

optimizer Contraction path optimizer. Check EinExprs documentation for more info.

outputs Indices that won't be contracted. Defaults to open indices.

kwargs Options to be passed to the optimizer.

See also: contract.

Missing docstring.

Missing docstring for contract(::Tenet.TensorNetwork). Check Documenter's build log for details.

Tenet.contract! — Function

contract!(tn::TensorNetwork, index)

In-place contraction of tensors connected to index.

See also: contract.