Introduction

In Tenet, we define a Quantum Tensor Network as a TensorNetwork with a notion of sites and directionality.

Tenet.Quantum — Type

Quantum <: Ansatz

Tensor Network Ansatz that has a notion of sites and directionality (input/output).

Tenet.plug — Function

plug(::TensorNetwork{<:Quantum})
plug(::Type{<:TensorNetwork})

Return the Plug type of the TensorNetwork. The following Plugs are defined in Tenet:

State Only outputs.

Operator Inputs and outputs.

Property No inputs nor outputs.

Tenet.sites — Function

sites(tn::TensorNetwork{<:Quantum})

Return the sites in which the TensorNetwork acts.

Tenet.tensors — Method

tensors(tn::TensorNetwork{<:Quantum}, site::Integer)

Return the Tensor connected to the TensorNetwork on site.

See also: sites.

Tenet.boundary — Function

boundary(::TensorNetwork)
boundary(::Type{<:TensorNetwork})

Return the Boundary type of the TensorNetwork. The following Boundarys are defined in Tenet:

Open

Periodic

Infinite

Adjoint

Base.adjoint — Function

adjoint(tn::TensorNetwork{<:Quantum})

Return the adjoint TensorNetwork.

Implementation details

The tensors are not transposed, just conj! is applied to them.

Concatenation

Base.hcat — Method

hcat(A::TensorNetwork{<:Quantum}, B::TensorNetwork{<:Quantum}...)::TensorNetwork{<:Composite}

Join TensorNetworks into one by matching sites.

Norm

LinearAlgebra.norm — Method

norm(ψ::TensorNetwork{<:Quantum}, p::Real=2)

Compute the $p$-norm of a Quantum TensorNetwork.

See also: normalize!.

LinearAlgebra.normalize! — Method

normalize!(ψ::TensorNetwork{<:Quantum}, p::Real = 2; insert::Union{Nothing,Int} = nothing)

In-place normalize the TensorNetwork.

Keyword Arguments

insert Choose the way the normalization is performed:

If insert=nothing (default), then all tensors are divided by $\sqrt[n]{\lVert \psi \rVert_p}$ where n is the number of tensors.

If insert isa Integer, then the tensor connected to the site pointed by insert is divided by the norm.

Both approaches are mathematically equivalent. Choose between them depending on the numerical properties.

See also: norm.

Fidelity

Tenet.fidelity — Function

fidelity(ψ,ϕ)

Compute the fidelity between states $\ket{\psi}$ and $\ket{\phi}$.